Autonomous enemies
I have made it so that the enemies in the Arraylist follow the shortest
path to Pac. This isn't finished yet because a) they all congregate on
the same point eventually b) I want them to only move at right angles to
Pac. At the moment sometimes they go diagonally